2023-01 upgrade: update pipeline, update to .net 7, fix directory structure, apply stylecop, remove codeql
