+++ README.md
@@ -7,9 +7,8 @@
 
 [![Github All Releases](https://img.shields.io/github/downloads/kasthack-labs/kasthack.Whitewashing/total.svg)](https://github.com/kasthack-labs/kasthack.Whitewashing/releases/latest)
 [![GitHub release](https://img.shields.io/github/release/kasthack-labs/kasthack.Whitewashing.svg)](https://github.com/kasthack-labs/kasthack.Whitewashing/releases/latest)
-[![license](https://img.shields.io/github/license/kasthack-labs/kasthack.Whitewashing.svg)](LICENSE)
+[![License](https://img.shields.io/github/license/kasthack-labs/kasthack.Whitewashing.svg)](LICENSE)
 [![.NET Status](https://github.com/kasthack-labs/kasthack.Whitewashing/workflows/.NET/badge.svg)](https://github.com/kasthack-labs/kasthack.Whitewashing/actions?query=workflow%3A.NET)
-[![CodeQL](https://github.com/kasthack-labs/kasthack.Whitewashing/workflows/CodeQL/badge.svg)](https://github.com/kasthack-labs/kasthack.Whitewashing/actions?query=workflow%3ACodeQL)
 [![Patreon pledges](https://img.shields.io/endpoint.svg?url=https%3A%2F%2Fshieldsio-patreon.vercel.app%2Fapi%3Fusername%3Dkasthack%26type%3Dpledges&style=flat)](https://patreon.com/kasthack)
 [![Patreon patrons](https://img.shields.io/endpoint.svg?url=https%3A%2F%2Fshieldsio-patreon.vercel.app%2Fapi%3Fusername%3Dkasthack%26type%3Dpatrons&style=flat)](https://patreon.com/kasthack)
