+++ src/.env
@@ -1,0 +1,4 @@
+#Bot token. Can be acquired at https://t.me/botfather
+ApostlePeter__Telegram__Token=
+ApostlePeter__Vk__ServiceToken=
+ApostlePeter__Vk__UserToken=
