--- src/Whitewashing/Whitewashing.Console/Configuration/BlackListOptions.cs
@@ -1,9 +1,0 @@
-﻿namespace Whitewashing.Console
-{
-
-    public class BlackListOptions
-    {
-        public string[] Words { get; set; }
-        public int[] Pages { get; set; }
-    }
-}
