--- src/Whitewashing/Whitewashing.Console/Configuration/Configuration.cs
@@ -1,9 +1,0 @@
-﻿namespace Whitewashing.Console
-{
-    public class Configuration
-    {
-        public TelegramOptions Telegram { get; set; }
-        public BlackListOptions BlackList { get; set; }
-        public VkOptions Vk { get; set; }
-    }
-}
