--- src/Whitewashing/Whitewashing.Console/Configuration/TelegramOptions.cs
@@ -1,7 +1,0 @@
-﻿namespace Whitewashing.Console
-{
-    public class TelegramOptions
-    {
-        public string Token { get; set; }
-    }
-}
