--- src/Whitewashing/Whitewashing.Console/Configuration/VkOptions.cs
@@ -1,11 +1,0 @@
-﻿namespace Whitewashing.Console
-{
-    public class VkOptions
-    {
-        public int AppId { get; set; }
-        public string Privileges { get; set; }
-        public string ServiceToken { get; set; }
-        public string UserToken { get; set; }
-        public string RedirectUri { get; set; }
-    }
-}
