--- src/Whitewashing/Whitewashing.Console/Models/Reports/AccountCheckResult.cs
@@ -1,5 +1,0 @@
-﻿namespace Whitewashing.Console.Models.Reports
-{
-    using System.Collections.Generic;
-    public record AccountCheckResult(IReadOnlyList<PostCheckResult> Posts, IReadOnlyList<GroupCheckResult> Groups);
-}
