--- src/Whitewashing/Whitewashing.Console/Models/Reports/GroupCheckResult.cs
@@ -1,4 +1,0 @@
-﻿namespace Whitewashing.Console.Models.Reports
-{
-    public record GroupCheckResult(bool Result, GroupIdentifier GroupId, string? Description);
-}
