--- src/Whitewashing/Whitewashing.Console/Models/Reports/GroupIdentifier.cs
@@ -1,7 +1,0 @@
-﻿namespace Whitewashing.Console.Models.Reports
-{
-    public record GroupIdentifier(int Id)
-    {
-        public override string ToString() => $"https://vk.com/club{this.Id}";
-    }
-}
