--- src/Whitewashing/Whitewashing.Console/Models/Reports/PostCheckResult.cs
@@ -1,4 +1,0 @@
-﻿namespace Whitewashing.Console.Models.Reports
-{
-    public record PostCheckResult(bool Result, PostIdentifier PostId, string? Description);
-}
