--- src/Whitewashing/Whitewashing.Console/appsettings.json
@@ -1,66 +1,0 @@
-{
-  "Whitewasher":
-  {
-    "Telegram": {                                       //| Telegram configuration
-      //"Token": "<set in user secrets>"                  | bot token. Can be acquired at https://t.me/botfather
-    },
-    "BlackList": {                                      // Detection configuration
-      "Words": [ // Blacklisted words
-        "навальн", "navalny",                           // Mentions of navalny & links to his team's social networks
-                                                        // "navalny.com",
-                                                        // "facebook.com/teamnavalny.spb",
-                                                        // "twitter.com/teamnavalny_spb",
-                                                        // "instagram.com/teamnavalny_spb",
-        "t.me/palace_square", "protestlivebot",         // links to protestor support bots
-        "fbkinfo", "fbk", "фбк",                        // fbk
-        "vbX30rvBEQM", "EiN4XgbKUfM"                    // Rammstein's video, https://tvrain.ru/news/eks_glavu_arhangelskogo_shtaba_navalnogo_prigovorili_k_25_godam_kolonii_za_repost_klipa_rammstein-529147/
-      ],
-      "Pages": [                                        //Blacklisted vk page ids
-        -55284725,  //teamnavalny
-        -155245615, //teamnavalny_eysk
-        -139246969, //teamnavalny_spb
-        -83214227,  //navalny.group
-        -150565101, //navalny_live
-        129244038   //navalny
-      ]
-    },
-    "Vk": { // VK service configuration
-      //"ServiceToken": "<set in user secrets>",        // Application service token, can be acquired at https://vk.com/editapp?&section=options&id=<your_app_id>
-      //"UserToken": "<set in user secrets>",           // User request token. OAuth URL can be generated with ConsoleTestService(Pass it to AddHostedService<T>() in Program.cs instead of BotService)
-      "AppId": 7838665,                                 // Your app id. Only used for OAuth URL generation.
-      "Privileges": "offline,wall,groups",              // App permissions. Only used for OAuth URL generation.
-      "RedirectUri": "https://oauth.vk.com/blank.html"  // Redirect URL for OAuth. Only used for OAuth URL generation.
-    },
-  },
-  "Serilog": {                                          // Logger configuration.
-    "MinimumLevel": "Debug",
-    "WriteTo": [
-      {                                                 // Colored console output
-        "Name": "Console",
-        "Args": {
-          "theme": "Serilog.Sinks.SystemConsole.Themes.SystemConsoleTheme::Literate, Serilog.Sinks.Console"
-        }
-      }
-
-      // Disk logging is disabled due to privacy requirements
-
-      //,
-      //{
-      //  "Name": "Logger",
-      //  "Args": {
-      //    "configureLogger": {
-      //      "WriteTo": [
-      //        {
-      //          "Name": "RollingFile",
-      //          "Args": {
-      //            "pathFormat": "log.{Date}.log",
-      //            "retainedFileCountLimit": 5
-      //          }
-      //        }
-      //      ]
-      //    }
-      //  }
-      //}
-    ]
-  }
-}
