+++ src/kasthack.ApostlePeter.sln
@@ -1,0 +1,33 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.5.33209.295
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "kasthack.ApostlePeter", "kasthack.ApostlePeter\kasthack.ApostlePeter.csproj", "{0078C44D-E87C-41E8-93AB-464A3A9AF6B0}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{32426B05-6BA5-4C61-AAE7-7B7037467EBE}"
+	ProjectSection(SolutionItems) = preProject
+		.editorconfig = .editorconfig
+		.env = .env
+		Directory.Build.props = Directory.Build.props
+		docker-compose.yml = docker-compose.yml
+	EndProjectSection
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{0078C44D-E87C-41E8-93AB-464A3A9AF6B0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0078C44D-E87C-41E8-93AB-464A3A9AF6B0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0078C44D-E87C-41E8-93AB-464A3A9AF6B0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0078C44D-E87C-41E8-93AB-464A3A9AF6B0}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {3652A3A5-215F-40A9-A918-8236F3A0C2F5}
+	EndGlobalSection
+EndGlobal
