+++ src/kasthack.ApostlePeter/Configuration/BlackListOptions.cs
@@ -1,0 +1,8 @@
+﻿namespace kasthack.ApostlePeter.Configuration;
+
+public class BlackListOptions
+{
+    public string[] Words { get; set; } = default!;
+
+    public int[] Pages { get; set; } = default!;
+}
