+++ src/kasthack.ApostlePeter/Configuration/TelegramOptions.cs
@@ -1,0 +1,6 @@
+﻿namespace kasthack.ApostlePeter.Configuration;
+
+public class TelegramOptions
+{
+    public string Token { get; set; } = default!;
+}
