+++ src/kasthack.ApostlePeter/Models/Reports/AccountCheckResult.cs
@@ -1,0 +1,5 @@
+﻿namespace kasthack.ApostlePeter.Models.Reports;
+
+using System.Collections.Generic;
+
+public record AccountCheckResult(IReadOnlyList<PostCheckResult> Posts, IReadOnlyList<GroupCheckResult> Groups);
