+++ src/kasthack.ApostlePeter/Models/Reports/GroupCheckResult.cs
@@ -1,0 +1,3 @@
+﻿namespace kasthack.ApostlePeter.Models.Reports;
+
+public record GroupCheckResult(bool Result, GroupIdentifier GroupId, string? Description);
