+++ src/kasthack.ApostlePeter/Models/Reports/GroupIdentifier.cs
@@ -1,0 +1,6 @@
+﻿namespace kasthack.ApostlePeter.Models.Reports;
+
+public record GroupIdentifier(int Id)
+{
+    public override string ToString() => $"https://vk.com/club{this.Id}";
+}
