+++ src/kasthack.ApostlePeter/Models/Reports/PostCheckResult.cs
@@ -1,0 +1,2 @@
+﻿namespace kasthack.ApostlePeter.Models.Reports;
+public record PostCheckResult(bool Result, PostIdentifier PostId, string? Description);
